Fix tests in CI. Also run p2p only once.
