+++ benches/src/db_lookup_times_utils/mod.rs
@@ -14,20 +14,16 @@ mod tests {
     };
     use fuel_core::state::rocks_db::RocksDb;
 
-    use crate::{
-        db_lookup_times_utils::seed::{
-            insert_compressed_block,
-            insert_full_block,
-        },
-        utils::ShallowTempDir,
+    use crate::db_lookup_times_utils::seed::{
+        insert_compressed_block,
+        insert_full_block,
     };
 
     const TEST_HEIGHT: u32 = 1;
     const TEST_TX_COUNT: u32 = 10;
 
     fn setup_test_db() -> RocksDb<BenchDatabase> {
-        let temp_dir = ShallowTempDir::new();
-        RocksDb::default_open(temp_dir.path(), None, -1).unwrap()
+        RocksDb::default_open_temp(None).unwrap()
     }
 
     #[test]
