+++ crates/fuel-core/src/executor.rs
@@ -3061,6 +3061,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(not(feature = "wasm-executor"))]
     fn block_producer_never_includes_more_than_max_tx_count_transactions_with_bad_tx_source(
     ) {
         let block_height = 1u32;
