+++ tests/Cargo.toml
@@ -28,6 +28,8 @@ cynic = { workspace = true }
 ethers = "2"
 fuel-core = { path = "../crates/fuel-core", default-features = false, features = [
   "smt",
+  "p2p",
+  "relayer",
   "wasm-executor",
   "test-helpers",
 ] }
@@ -43,7 +45,7 @@ fuel-core-p2p = { path = "../crates/services/p2p", features = [
 fuel-core-poa = { path = "../crates/services/consensus_module/poa" }
 fuel-core-relayer = { path = "../crates/services/relayer", features = [
   "test-helpers",
-], optional = true }
+] }
 fuel-core-storage = { path = "../crates/storage", features = ["test-helpers"] }
 fuel-core-trace = { path = "../crates/trace" }
 fuel-core-txpool = { path = "../crates/services/txpool_v2", features = [
@@ -79,8 +81,6 @@ proptest = { workspace = true }
 tracing = { workspace = true }
 
 [features]
-default = ["fuel-core/default", "relayer"]
-p2p = ["fuel-core/p2p", "fuel-core-p2p"]
-relayer = ["fuel-core/relayer", "fuel-core-relayer"]
-wasm-executor = ["fuel-core/wasm-executor"]
+default = ["fuel-core/default"]
+only-p2p = ["fuel-core-p2p"]
 aws-kms = ["dep:aws-config", "dep:aws-sdk-kms", "fuel-core-bin/aws-kms"]
